Solucionado errores en 'repositories'

+++ src/main/kotlin/repositories/EntrenadorRepository.kt
@@ -1,6 +1,5 @@
 package org.example.repositories
 
 import org.example.models.Entrenador
-import org.example.repository.CrudRepository
 
-interface EntrenadorRepository : CrudRepository<Entrenador, Int>
+interface EntrenadorRepository : CrudRepository<Entrenador, Long>
